requirements.txt
flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.31.0